requirements.txt
langchain 
langchain_experimental 
langchain_huggingface 
chromadb 
langchain_community
langchain-openai
python-dotenv
docling